Update to Rails 2.3.5
Rails 2.3.5 was rushed out because of a possible XSS vulnerability in
the "strip_tags" method for pages viewed using Internet Explorer.

This application only uses "strip_tags" in one place (when displaying
"tweet" titles), and the input is always trusted seeing as only the
admin can author tweets.

So there is no need to perform the upgrade for security; instead, I'm
doing the upgrade now so as to catch any potential regressions in the
release sooner rather than later, seeing as so many other changes are
bundled along with the security fix:

  $ git diff --shortstat v2.3.4..v2.3.5
  181 files changed, 2267 insertions(+), 823 deletions(-)
  $ git oneline v2.3.4..v2.3.5 | wc -l
       131

Signed-off-by: Wincent Colaiuta <win@wincent.com>