Set "envelope from" on all outgoing email
In the previous commit we switch the delivery method in production mode
from the default (:smtp) to :sendmail, so that we could take advantage
of the automated forwarding we have set up on the server through
authsmtp.com.

The problem with this set-up is that the "envelope from" header:

  Return-Path: <rails.wincent.com@wincent1.inetu.net>
  X-Envelope-From: rails.wincent.com@wincent1.inetu.net
  From: win@wincent.com

doesn't match the "From" header. This may cause some spam filters to
complain.

So, we explicitly pass in a "return-path" header, which according to the
Rails API documentation, is used to set the "envelope from" address.
Presumably, Rails is using the "-f" switch to sendmail behind the scenes
to achieve this, in addition to the default switches ("-i -t").

In order to make the "-f" switch work, the Rails user running the app
must appear in the /etc/mail/trusted-users file.

Signed-off-by: Wincent Colaiuta <win@wincent.com>