Return 403 status for attempts to comment when commenting is disabled
Since adding exception notifications I've been getting a steady stream
of emails like this:

          Exception: RuntimeError
  Controller#action: comments#new
                URL: https://wincent.com/twitter/10/comments/new
            Message: parent does not accept comments

The cause is evidently spammers who are systematically generating URLs
in the hope of finding spammable forms, or perhaps overzealous search
engines which having found similar pages for tweets where commentability
is turned on think there might be something at the new URL.

As this is not really an internal server error, a 403 error code and no
exception email is what we really want here.

There are a few other places in the code base that currently just do a:

  raise "something isn't right here"

Which could similarly be tweaked to instead return 403s (or 404s, or
whatever is suitable) but I'll handle those when I start to see
exception notifications for them coming in.

Signed-off-by: Wincent Colaiuta <win@wincent.com>