Add "meta/whitespace" specs
We check the codebase for three things:

  - trailing whitespace
  - tabs
  - missing newlines on last line of files

We don't check vendor code (anything in "vendor"), instead just
focussing on the relevant parts of "app", "config", "lib", "spec" and so
on.

We specify some manual exclusions for some files created by Cucumber
which have whitespace damage in them, and which will just get recreated
on each Cucumber update so there is no point in fixing the damage.

Stuff that won't get recreated with each update is whitespace-fixed (for
example, 5 files inside "config/initializers"), as well as some files of
my own creation with whitespace errors in them.

Signed-off-by: Wincent Colaiuta <win@wincent.com>