deploy: point at new server
Note that this is only the first step. Now that we are on EC2 our
staging environment will no longer be on the same machine and in fact we
can't even know what its IP is until runtime.

Likewise, our Git repos are no longer on the same machine so that will
require some set-up as well.

Signed-off-by: Wincent Colaiuta <win@wincent.com>