deploy: drop check for rubygems-update
As these are clean installs there is no rubygems-update gem on the
system. At the moment my general instance is running 1.3.6 and the Rails
instance is running 1.3.5, and as future instances will be based on
these snapshots there's really no point in keeping the (now old) check
for the 1.3.1 rubygems-update.

Signed-off-by: Wincent Colaiuta <win@wincent.com>