Fix more breakage in issues/_search partial view spec
Commit 254006f fixed some breakage due to an error in the template. Once
that error was fixed, execution was able to continue on, thus revealing
this new error, this time in the spec.

The problem is that only products where "hide_from_front_page" is set to
false appear in the menu, but the setting is actually true by default so
the created products in the test didn't actually appear.

Signed-off-by: Wincent Colaiuta <win@wincent.com>