products.js: set width/height using attr()
If we use the new jQuery (1.4) element creation syntax to pass in width
and height, these end up getting defined as CSS packed into a "style"
attribute on the element.

This isn't what we want.

By separately chaining an attr() call and using that to set the width
and height we get real width and height attributes, which is what we
wanted.

Signed-off-by: Wincent Colaiuta <win@wincent.com>