Allow hyphens in forum names
This requires us to add a "permalink" field to the database because
while it is very easy to go from "name" to "param" (ie. "Foo Bar" to
"foo-bar") it is not so easy to go from "param" to "name" for the
purposes of finding the forum in the database (eg. is "foo-bar"
equivalent to "Foo Bar" or to "Foo-Bar"?).

See:

  https://wincent.com/issues/1506

Signed-off-by: Wincent Colaiuta <win@wincent.com>