Rename SearchController to SearchesController
Seeing as the change to the Rails 3 routing DSL is a pretty big one, we
may as well take the opportunity to iron out some inconsistencies in the
existing set-up.

We are "abusing" the search controller by making it act like a resource
when it isn't a "real" resource with a database-backed model.

So let's rename it in the plural to at least make it consistent with
other "real" resource controllers, and tighten up the routing as well.

One day might record searches in which case searches#index will become a
(potentially admin-only) view of recent searches on the site. (I say
admin-only because otherwise people could use it to post spam links and
other rubbish without having to wait for moderation.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>