Rationalize issue search routing
It was always a mistake to have a special "issues" action on the
SearchController. It makes much more sense to have a "search" action on
the IssuesController and in the URLs and view templates have "issues" as
the scope in which the "search" actions reside (both initial display of
the search form and display of the results list).

This commit organizes things much more nicely, giving us the following:

  - GET /search is not routable
  - GET /search/new maps to searches#new and displays the site search
    form
  - POST /search maps to searches#create and runs an site search and
    displays the result
  - GET /issues/search shows the issue search form via issues#search
  - POST /issues/search does the search and shows the results (via
    issues#search)
  - issue search templates are now stored under:
    - app/views/issues/search/new (for new searches)
    - app/views/issues/search/create (for results)

Signed-off-by: Wincent Colaiuta <win@wincent.com>