Teach relative date JS to grok UTC dates
In the move to Rails 3 date strings returned from models have changed
from:

  Mon Mar 01 04:22:02 -0500 2010

To:

  2010-04-14 23:50:46 UTC

As JavaScript chokes on the latter, we have to add an explicit
dateFromUTCString() function that takes such a string and converts it
into a Date object.

At the same time, given that UTC dates are a little bit unfriendly to
read, we switch to using the toLocaleDateString() and toLocaleString()
methods for populating the tooltip and "distant" dates.

Signed-off-by: Wincent Colaiuta <win@wincent.com>