bugfix: the JS method Date.UTC expects zero-based months
Commit 72b98ec introduced buggy code for getting Date object from a UTC
string. The Date.UTC method expects zero-based months (0-11) and we were
passing in one-based months (1-12); blame my O'Reilly JavaScript pocket
reference for not mentioning it.

This bug cause some weird looking date values, so in that same commit I
ended up commiting another bug as well, reversing the order of the "now"
and "then" operands in the difference calculation to give me dates which
looked at first glance to be more "reasonable".

On doing additional testing, and now with a wider spread of sample
dates, it was easy to see both the initial bug and the misguided "fix"
which came with it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>