Haml: explicitly set up to emit HTML5
It appears that Haml is already automatically emitting HTML5 under Rails
3, but be explicit about it anyway, by dropping the

  <?xml version="1.0" encoding="UTF-8"?>

tag, which is not needed in normal (text/html) HTML5 documents, and also
switching the Haml "format" option to :html5.

See

  http://dev.w3.org/html5/html4-differences/

for info on the differences between HTML4 and HTML5, as well as the XML
version of HTML5 documents.

Signed-off-by: Wincent Colaiuta <win@wincent.com>