Make GET /search work
We model searches as RESTful resource, so we present a search form for
/search/new and actually run the search and display the results by
POSTing to the "create" action of the SearchesController.

Nevertheless a user that knows that the site offers search may try to
guess the search form URL as /search, and we may as well make it work
for them.

Signed-off-by: Wincent Colaiuta <win@wincent.com>