Serve dynamic JS with appropriate Content-Type
Previously we requested dynamic JS via URLs like:

  /js/issues/show

Even though our controller sends JS in response, the browser console
ends up reporting an error about receiving JS with a "text/html"
Content-Type.

So we change our URLs to:

  /js/issues/show.js

And the browser stops warning.

(Note that the warning was really cosmetic only, as the system still
worked fine, but it is nicer to keep things warning-free.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>