Replace some html_safe calls with raw()
While the behavior is identical (raw just calls html_safe) this is a
little more faithful to what's actually going on here.

html_safe is the right method to use in helpers (ie. we're marking
something as "safe to use" so that the view can use it without having to
take any further action).

But in views raw is usually the right choice. It does the same thing but
better describes what is actually happening ie. it says "emit this raw
HTML without escaping it" rather than "I know this content to be safe".

As an added bonus this also saves us a few characters.

Signed-off-by: Wincent Colaiuta <win@wincent.com>