JsController: return 404 for non-existent templates
Previously a request for a non-existent template would cause an HTTP 500
status code (internal server error due to do ActionView::MissingTemplate
exception).

We don't want to produce exception notifications for this kind of thing,
so instead check for the existence of the template and return a 404 if
not found.

We have 3 alternatives here for how to render the 404.

  (1) render :file => "#{Rails.root}/public/404.html", :status => 404

  (2) raise ActiveRecord::RecordNotFound

  (3) render :text => '', :status => 404

Here the third option is the most appropriate, just like it is for
missing feeds, because this controller should supposedly only be serving
legitimate machine-originated requests (ie. other pages on the site
requesting JavaScript) rather than end-user-originated requests.

For this reason the second option is completely inappropriate (it ends
up redirecting to the main index and showing a flash error), and the
first option is better but still goes too far (no need to show an entire
404 error page).

Tested using:

  curl -k -i -H 'Accept: */*' https://localhost:3002/js/issues/create.js

Signed-off-by: Wincent Colaiuta <win@wincent.com>