Add error messages partial
This was previously factored out because Rails 3 doesn't provide the
"error_messages_for" method any more (it was ripped out into a plug-in).

Rather than installing such a plug-in, it is vastly preferrable to just
stick things in a partial for easy customization.

Note that the original version of the partial assumed a non-nil "model"
variable, but in this commit a guard for that has been added to prevent
possible exceptions.

(For example, in our links#new action we don't bother creating a dummy
Link instance when first showing the form, so the "model" variable will
be nil and trying to call "model.errors" on it would cause an
exception.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>