Simplify links#new template
Creating a blank/new record in our controller action allows us to use
the recommend resource-based pattern with "form_for" (ie. we just pass
in the model and Rails derives the form URL via introspection).

Eventually, if we do this across the entire application, we can change
our error messages partial to drop the nil check, because we will always
have a non-nil model available to the partial.

Signed-off-by: Wincent Colaiuta <win@wincent.com>