Form shakedown
This is the first step in a rationalization of all forms in the app.

We introduce ValidatingFormBuilder, which at the moment does nothing but
will later insert inline validation error messages directly into the
form itself.

And we dump the table as the default method of form organisation.
Instead switch to labels, and we'll worry about making the CSS force
things to line up later on.

Note that we don't use an explicit "for" attribute in the label element.
Instead we put the input field directly inside the label, which is
supported in all sensible browsers and is easier to maintain, as we
don't have to worry about the exact ID that Rails will assign to the
input element.

For now the only complication is that during validation, what would
normally display as:

  <label>
    URI
    <input ...>
  </label>

Will display as:

  <label>
    URI
    <div class="field_with_errors">
      <input ...>
    </div>
  </label>

So our CSS will have to ensure consistent display both with and without
that internal "div" element. We may have to abuse the
ValidatingFormBuilder class to wrap even fields without errors in some
kind of "div" for consistency, although I would prefer to avoid that if
possible.

Signed-off-by: Wincent Colaiuta <win@wincent.com>