Use lists in forms for prettier CSS
Strictly speaking this is an "abuse" of lists, but we can rationalize it
because the form contains a "list" of input fields.

Less ugly than using tables for layout, and buys us a nice shortcut when
it comes to CSS styling. For example, we can draw a divider line
underneath each "row" by styling the "li" element, something that would
be next to impossible with separate "label" and "input" elements with no
containing element grouping them together.

Still to be resolved is the display of validation errors, because the
input field winds up in a "div" element whenever a validation error
occurs.

Signed-off-by: Wincent Colaiuta <win@wincent.com>