Dump ValidatingFormBuilder in favor or field_error_proc
This fixes an annoying wart in the Rails form handling; namely, that
fields with validation errors are wrapped in "div" tags of class
"field_with_errors" rather than simply adding the "field_with_errors"
class to the existing element.

This is annoying because it makes CSS stying of forms hard (valid
models have one DOM hierarchy and invalid models have another). In a
worst case scenario the "div" element might even get inserted into the
DOM at a point where it's not valid for such an element to appear.

So, we replace the annoying default by supplying our own
"field_error_proc".

Unfortunately, it is not entirely pretty. The way we do this is by
scanning the HTML "tag" for an existing "class" attribute; if found we
add our error class to the attribute, and if not we add the attribute.

This is a bit brittle as it depends on Rails always supplying us with a
valid tag, but it seems to be a necessary evil.

So we use a couple of extremely conservative regexes to minimize the
risk of breakage. Note how although the Rails code refers to the first
parameter as a "tag", it could actually be any string of HTML; eg.

  <label for="foo_attribute">
    Arbitrary HTML in here, including other nested elements...
    Oh, and look, <span class="gotcha">it's a trap!</span>
  </label>

So we can't just blindly scan for 'class=...' but must instead look for
it only in the initial element anchored at the start of the string.

On the bright side, this field_error_proc could turn out to be just the
mechanism we were looking for earlier in our search for the best means
of in-lining validation error information via form tool-tips.

Signed-off-by: Wincent Colaiuta <win@wincent.com>