Make button/submit inputs match size of "link" CSS
For some reason, both Safari and Firefox use a smaller font for "input"
elements of type "submit" and "button" (11px) than they use for "a"
elements wrapped in the "link" class (12px).

Force them to be the same size by explicitly specifying a font-size of
12px for both.

This makes the "buttons" produced by the "link" class and the real
buttons produced by "input" elements almost identical in size.

Signed-off-by: Wincent Colaiuta <win@wincent.com>