Dump Cucumber in favor of Steak
The one thing that was holding me back from doing this earlier is that I
didn't know how to replicate the "tags" functionality that I use from
Cucumber in order to switch Capybara drivers for JavaScript and
non-JavaScript scenaros.

But the answer lies here:

  http://gist.github.com/397977

So this commit rips out Cucumber and replaces it with Steak. Note that
the "features" directory is still present, and won't be removed until
those Cucumber features have been migrated over to Steak.

Signed-off-by: Wincent Colaiuta <win@wincent.com>