More updates for Capybara/Steak
This gets us a sample spec which would work if it weren't for the
strictness of the Rails router.

(An HTTP request for "/" will actually raise a RoutingError, because our
routes insist on HTTPS-only access. It seems that when the request goes
through rack_test we have the same problem even though there is no
actual network transport involved.)

Will have to think about our options here: either hacking
Capybara/Rails/Rack to fake or do an HTTPS request, or relaxing the
routing rules (which might cause problems with our route generation).

Signed-off-by: Wincent Colaiuta <win@wincent.com>