Drop protocol requirement from routes.rb
I love forcing all traffic to be via HTTPS but the strict requirement at
the routing level is causing real problems when testing.

For example, a request for http://localhost:3000/ will raise
RoutingError because no route matches "/" over HTTP.

This causes problems for Steak/Capybara, even when using the rack_test
driver. So, in the interests of simplicity drop the requirement.

This will mean that we have to check very carefully at our application
boundaries to make sure we publish always the right protocol and host;
more than anything this means:

  - links in outgoing mails
  - links in feeds
  - links in anything that might get cached (and which other users will
    see)

Will need to beef up our tests and possibly even make some helpers to
make enforcing this policy easier.

Signed-off-by: Wincent Colaiuta <win@wincent.com>