Add factories for Email, Product and User models
Note that we make this useful in the development environment by
providing randomized attributes (so that records can be created on the
fly for experimentation purposes without running into uniqueness
constraints).

In the test environment we use repeatable sequences instead, because we
want all test runs to be identical and we don't have to worry about
uniqueness because the database is truncated after each session.

Signed-off-by: Wincent Colaiuta <win@wincent.com>