Add "local_request?" method
This is needed in order to decide whether to send secure cookies or not.
If we send secure cookies during testing (which occurs over HTTP, not
HTTPS) we can't test any app functionality that requires the visitor to
be logged in. So for local requests we disable secure cookies.

This method replaces the "local_request?" method which used to be part
of Rails but was ripped out for Rails 3.

Signed-off-by: Wincent Colaiuta <win@wincent.com>