Make products category default to ''
Previously the default was NULL, so it was possible, at least from the
console, to create a product with a NULL category.

This led to a broken pop-up menu, because we could wind up with:

  null
    product1

    product2
  Consumer
    product3

(ie. three categories, "null", "", and "Consumer)

So set the default to "" and disallow NULL values. In my production
deployment there are no products with "" or NULL as their category, but
this is still a nice clean-up for the local development environment.

Signed-off-by: Wincent Colaiuta <win@wincent.com>