Don't explode if post title is nil
Never will happen during legitimate form submission (where even an empty
title field will still be passed in as an empty string), but in the
console we were exploding given:

  Post.create

because we ended up trying to call downcase() on the nil title. Better
to avoid the exception and let the failure kick in at validation time.

Signed-off-by: Wincent Colaiuta <win@wincent.com>