Fix broken tagging
Something is different in Rails 3. This previously worked in Rails 2:

  some_model.tags << tag_instance

Under Rails 3 this produces a warning about being unable to assign to
mass-protected attributes (tag_id, taggable_type, taggable_id). This in
turn means the necessary Tagging object can't be created because the
necessary fields are NULL as they were never assigned to.

So make those attributes mass-assignable, and hope that all other models
which have_many taggings do the right thing and don't all
mass-assignment to their taggings association.

Signed-off-by: Wincent Colaiuta <win@wincent.com>