Fail nicely for "forbidden" errors
ActionController::StatusCodes::STATUS_CODES was ripped out.

Replace by Rack::Utils::HTTP_STATUS_CODES.

This saves us from spewing ugly HTML into the AJAX error flash when we
do bad things (like trying to post a comment on a tweet for which
commenting is disabled).

Prior to this commit we spewed out a backtrace; now the flash just shows
the word "Forbidden" as it should.

(Still to do is suppressing the display of the comment form in the case
that commenting is disallowed.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>