Clear AJAX flash before starting new request
Slightly better than the last commit, remove any old error messages
before starting the request.

The way we were doing it before (removing the old message just before
displaying the new one) may happen too quickly for the user to see it
and actually realize that it is a new error, especially if the error is
the same in both cases.

Signed-off-by: Wincent Colaiuta <win@wincent.com>