Fix commenting on wiki articles with funny characters
The problem here was with wiki articles with titles like:

  bad <stuff> in this title

We were unpacking the URL components and getting:

  bad_%3Cstuff%3E_in_this_title

Causing our "find" to fail. We really needed to use the param provided
by Rails instead, which is appropriately escaped:

  bad_<stuff>_in_this_title

This bug only occurred for the wiki and not the other commentables
because the others all have sanitized paths. (Even the Post model has
relatively sanitized paths; the same title on a Post instance would
become something like "bad-stuff-in-this-title" as a permalink.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>