Use button_to instead of link_to for issues#show destroy
This is better thank link_to because it will still work even when
JavaScript is disabled (although we won't see a confirmation dialog),
and also because the set-up is simpler (no need to include the CSRF
protection token in the "head", nor worry about such a token ever
getting inadvertantly cached in the page).

Note how this change requires some additional CSS because Rails returns
a fairly hideous string of HTML consisting of a "form" element
containing a "div" which in turn contains the "input" butotn.

In a future commit I will probably add an intermediate confirmation page
that can be shown when JavaScript is not available.

Signed-off-by: Wincent Colaiuta <win@wincent.com>