Handle spec dependencies the right way
Prior to RSpec 2 we could include our shared example groups multiple
times and it would not complain.

In RSpec 2 this is not longer the case.

The "right" way to handle this then is to stick the shared example
groups in "spec/support", where they will be automatically required by
the spec_helper. We therefore remove all explicit "requires" of the
shared example files.

To ensure that they only get evaluated once, we need to make sure that
"spec_helper.rb" only gets evaluated once.

The easy way to do this would be to set up a once-only guard inside the
spec_helper itself, but the "right" way is to always require the helper
using an absolute path; then Ruby will automatically only evaluate it
the first time through, and ignore any subsequent requires for it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>