Add protocol to default_url_options
This is the second part of our treatment of issue #1563:

  https://wincent.com/issues/1563

We want to make sure that outgoing mails use the right host, protocol
and port (although in practice on the production server the port will
never be explicitly included).

Note that we do not set options if they are not set in APP_CONFIG because
we don't want nil values in the hash to trigger code paths like this one
in Rails (from actionpack/lib/action_dispatch/routing/route_set.rb):

  rewritten_url << ":#{options.delete(:port)}" if options.key?(:port)

(This would add ":" to the URL whenever the port option was not set in
the APP_CONFIG.)

Also note that we do not attempt to guard against user stupidity, like
setting "port" to 80 and "protocol" to "https", or "port" to 443 and
"protocol" to "http".

Signed-off-by: Wincent Colaiuta <win@wincent.com>