Change order of columns in issue lists
Give more importance to issue number and issue summary by making them
the left-most columns.

The next most interesting columns are product, kind, and status, in that
order.

Finally, in the "meta" category (stuff that you might want to glance at
but which is not fundamental) we have the timestamp and public/private
icon.

Not only does this sort things according to their interest/importance,
it also moves a lot of clickable elements that the user may not want to
click on further to the right.

In other words, a first-time visitor wanting to see a ticket is now more
likely to hit the right link (the summary) rather than erring and
hitting the "bug" or "open" links, which would take him to another list
of issues rather than the issue he wants to see.

See:

  https://wincent.com/issues/1477

Signed-off-by: Wincent Colaiuta <win@wincent.com>