Right align "labels" in issues#show template
In order to target these elements we group things using "thead" and
"tbody"; this allows us to distinguish between tables with the header
elements on the left (like in issues#show) and tables with the header
along the top (like in issues#index).

Note the secondary benefit here: in our issues#index we now get a
noticeable visual separation between header and the rest of the table.

Signed-off-by: Wincent Colaiuta <win@wincent.com>