Clean up email#edit template
Drop the table and use unordered list instead. Note how we move the
"User" info elsewhere on the page seeing as it doesn't really fit within
the form: the link back to the user and the user name itself are now in
the (new) breadcrumbs, and the "edit" link becomes a button down the
bottom.

Signed-off-by: Wincent Colaiuta <win@wincent.com>