Enhancements to pages forms
Move the submit button into the shared partial so that we can get rid of
the old table-based layout and use a list instead.

Note how we don't need to pass a parameter to the "submit" method; Rails
will give it a title of "Create Page" or "Update Page" depending on
whether the record associated with the form is a new record.

Signed-off-by: Wincent Colaiuta <win@wincent.com>