DRY up stylesheet inclusion
Use the new "uses_stylesheet_links" method to DRY up views.

Note that in some cases we still need an explicit "stylesheet_link_tag"
because we can't configure that from within the controller; for example,
the "support" controller actually shows issues, but if we specify
"uses_stylesheet_links" in the support controller, it will include
"support.css" when what we want is "issues.css".

The pages and products controllers are an interesting exception: we do
show pages, but always through the products controller, so there a
simple "uses_stylesheet_links" will suffice (it includes
"products.css"). This is probably the only place in the application
where a nested resource is actually handled by its parent controller
(which may itself be a design smell, but may also be an unavoidable
consequence of the desire to have clean URLs without a "pages" component
in them).

Signed-off-by: Wincent Colaiuta <win@wincent.com>