Convert issue comments to "bubbles"
This is an intermediate step towards solving the problem described in
ticket #1530:

  https://wincent.com/issues/1530

In short, the table-based comment layout makes it impossible to have
scaling "pre" elements and we have to use fixed-width ones which is not
optimal.

The next step will be to use the "boxed" class to turn these speech
bubbles back into rectangular boxes resembling tables.

Signed-off-by: Wincent Colaiuta <win@wincent.com>