First cut at "boxed" comment style
This displays correctly in Safari but is still only a rough cut as it
suffers from alignment problems when the text is zoomed in or zoomed
out by a large degree.

The placement of the "edit" link is also a little ugly, and the font
size for the author is probably to large.

A future refinement will likely be using 0px margins for both inner
elements (the "blockquote" and the "cite") and applying the border
and background to the "li" container itself. In this way even when
resizing to extreme dimensions the borders can't possibly become
misaligned.

Also, may move the "bubbled" and "boxed" styles into separate files so
that they can be easily imported into the desired controller-specific
CSS files rather than the site-wide "display.css" file.

Signed-off-by: Wincent Colaiuta <win@wincent.com>