Intermediate cut at "boxed" comment style
This is (obviously) not intended for deployment as is, but it is an
intermediate step in the development of the boxed comment style.

It looks like we have no choice but to use floated elements for now, and
as those are taken out of the document flow, the containing element (the
"li") collapses to zero height because it has nothing in it.

So we have to introduce the clearfix hack/kludge (an additional div with
no semantic meaning in the HTML).

I tried adding this in with the ":after" pseudo-selector but it did not
work, so a real element is required.

This still doesn't give us what we want though because as you can see,
on being removed from the flow our width specifications of 80%/20% now
mean "80%/20%" of the window width, which now makes them too wide.
Specifying a fixed width will defeat the purpose of this change, as the
whole point is to make out comments flexible width and have flexible
"pre" elements inside them.

The other strategy I tried was making the "li" elements using relative
positioning, and then the contained "blockquote" and "cite" elements use
absolute positioning, to put them in the top left and top right corners
respectively.

This "worked" too, but had the same problems (the containing element
collapsed away to nohting, and the width could not longer be meaninfully
specified in terms of a percentage).

So my next strategy is going to be to see if I can learn from what I did
with the form CSS, which also consists of lists with contained elements,
and which exhibits the desired behaviour (flexible width, with no
"clearfix" hacks needed).

Signed-off-by: Wincent Colaiuta <win@wincent.com>