inline-block to the rescue
This is now close to acceptable. We reverse the order so that the "cite"
element comes before the "blockquote" element, in order to force the
former to appear on the left (seeing as we can neither use floats nor
relative positioning for the reasons already noted previously).

And we use "inline-block" display to make the two elements appear next
to each other.

The only "hacky" element right now is the manual fine-tuning we have to
do on the margins to make them coincide: because we don't know whether
the "cite" or the "blockquote" element will be taller, we have to make
both of them draw a border along their shared (inner) edge, otherwise
the dividing line won't fill the entire height of the containing box.

So we have to bump one 2px towards the other, and the other 3px towards
the other. (Well, I could have bumped just one, I guess, but doesn't
really matter.) They happen to line up perfectly in both Firefox and
Safari at all window sizes except really narrow ones, in which the two
elements wrap and no longer appear next to each other. Not a lot that we
can do about that.

Signed-off-by: Wincent Colaiuta <win@wincent.com>