Correct text darkness in boxed comments
They were turning out too light, because they were inheriting the
"lighter blockquote" styling that is used elsewhere in the site.

Likewise, lists were turning out too light.

So, we make "p" elements darker to compensate, and "li" elements too. We
let "blockquote" elements retain the lightenin behavior, because in this
way nested blockquotes appear lighter, which is nice when quoting
somebody to not give too much emphasis to the quoted text.

Signed-off-by: Wincent Colaiuta <win@wincent.com>