Remove hacky margin fine-tuning
Rather than trying to make the shared borders line up, set the border
only on one side, the blockquote side (because it is more likely to be
taller).

Make the "cite" element a little wider so that it is less likely to wrap
and become taller than the "blockquote" element.

And set a min-height of 80px (100% won't work because the parent
container, the "li" element, does not have a fixed/known height) so that
at reasonable window sizes the "blockquote" will always be as tall as
the "cite", even when there is only one line in the comment.

This is still not perfect, as at large font sizes and small window sizes
it is possible to make the containing box grow too tall and the
"blockquote" border will not extend to cover the full distance, but it
is close to being perfect.

Signed-off-by: Wincent Colaiuta <win@wincent.com>