Compromise on tweet hoverlink behavior
Try to get the best of both worlds.

On the tweets#index page we have the new behavior (hidden hoverlinks)
with the same layout as always (left-aligned info).

On the blog#index page in the sidebar we keep the old behaviour (always
visible permalinks) so that we can preserve the old layout
(right-aligned info).

Signed-off-by: Wincent Colaiuta <win@wincent.com>