Set up explicit BUNDLE_PATH for deployments
We want BUNDLE_PATH to be set to something that the Rails app user can
write to without needing to use "sudo", for security, isolation, and
ease of deployment.

We can't just set BUNDLE_PATH in ~/.bash_profile, even though we are
using "sh -l" in our deployment script, because the file won't be read
(if you look at "man sh" you'll see that Bash will only read the first
file it finds, which in our case is going to be /etc/profile.

So, to avoid explicitly passing in BUNDLE_PATH in our deployment script
(with "env BUNDLE_PATH=foo" or similar) we set it up in our
.bundle/config file at the project root.

Signed-off-by: Wincent Colaiuta <win@wincent.com>