Use mail method in comment mailer
It already works without this, but this represents "the way" to set up
mailers in Rails 3 (all the documentation suggests this).

The reason it works anyway is that when you send:

  CommentMailer.new_comment_alert

Note that you are sending a class method which doesn't exist. Rails
catches this, calls the instance method, and creates the Mail instance
if you didn't do it explicitly.

Best to create it explicitly, though, because it seems that the
"default" settings don't apply if you rely on Rails to create the
instance.

Signed-off-by: Wincent Colaiuta <win@wincent.com>