Further simplification to issue mailer
While the documentation says:

  If you need other headers not listed above, use the headers['name'] =
  value method.

Looking at the source code in action_mailer/base.rb reveals that the
mail method ends up just calling "header[foo] = bar" for any options
passed in anyway.

There is even a test in base_test.rb for this ("can pass random headers
in as a hash to mail").

So use this as an opportunity to clean up our source code, and trust the
specs to notify us of whether anything breaks because Rails changes the
way it works in the future.

In the meantime, have submitted a documentation patch upstream:

  https://rails.lighthouseapp.com/projects/8994/tickets/5007

Signed-off-by: Wincent Colaiuta <win@wincent.com>