Revert 'Remove "email_address" attribute from reset factory'
Turns out that if we want "Reset.make!" to return a valid instance we
need to provide this attribute after all.

This is because in order to create the instance, Factory Girl updates
the attributes and then calls "save", thus triggering the "update"
validation and tripping us up.

This may in fact be a signal of code smell, but I am not sure. I would
think that one should be able to do this to create a "valid" record, and
if it is valid before "save", then calling "save" should not cause it to
suddenly become not valid:

  r = Reset.new
  r.user = user
  r.email = user.emails.first
  r.valid? # => true
  r.save
  r.valid? # => false

This reverts commit a3af47f9.

Signed-off-by: Wincent Colaiuta <win@wincent.com>