Add basis for exception notifying mechanism
At this stage this is just an empty middleware which doesn't do
anything. This is the first step in replacing the old exception
notification mechanism (which was just an override of
rescue_action_in_public in the application controller).

The old method has no effect in Rails 3 because the
rescue_action_in_public method was extracted out into a separate
middleware in Rails commit 11af089cee0a0e744e267d32becfe2c66a586d31.

We therefore need to create our own middleware in order to catch
exceptions.

Signed-off-by: Wincent Colaiuta <win@wincent.com>