Fully armed and operational ExceptionReporter
Make sure that "controller" and "request" information is available to
the exception mailer.

Note that we even handle the case where the controller is not known,
although in practice this is unlikely to come up, I believe (if the
controller is not known, the error should actually be a RoutingError,
which won't trigger a notification).

Signed-off-by: Wincent Colaiuta <win@wincent.com>